gst-convert: Handle ==> selector in rewrite rules
When converting PetitParser client code from Pharo to GNU Smalltalk
one will need to replace longer binary selectors like ==> with a shorter
version of them. Use the parser class of the input to parse the rule.
Ideally we would use the parser for the left side of the rule and the
exporter for the right side of the rule but this is not possible yet.
Collect all rules until after the parser class was selected, then add
the rules to the converter.

This patch allows to parse a rule like this '(``@object ==> ``@arg1 )
-> (``@object => ``@arg1)'.

2012-09-29  Holger Freyther  <holger@freyther.de>

	* scripts/Convert.st: Collect all rewrite expressions and parse
	them with the parser for the code.