tests: Correct the Dates test and update the test result.
The test intended to verify that the suffix is not consumed from
the stream but the date was parsed from a new stream so the test
always succeeded. Right now more characters are consumed and the
suffix test would fail. Change the test to use the right stream
for parsing the date and change the test result to include the
suffix and what remained inside the stream.

2013-01-17  Holger Freyther  <holger@freyther.de>

	* tests/dates.ok: Fix the testcase to read from the right
	ReadStream.
	* tests/dates.st: Update the test result.