stinst: Allow gst-convert to work for OldSyntax/Squeak syntax
The gst-convert command

$ gst-convert -r'Osmo.LogManager -> LogManager' -F squeak -f gst

was rewriting the nodes but the exported file contained the original
unmodified sourcecode. The FormattingExporter for the gst syntax is
using the >>#methodFormattedSourceString to format, use it inside the
OldSyntaxExporter as well. Another option would be to subclass the
SqueakSyntaxExporter and OldSyntaxExporter and add the formatting there.

2013-02-10  Holger Hans Peter Freyther  <holger@freyther.de>

	* OldSyntaxExporter.st: Reformat the method node in
	OldSyntaxExporter>>#oldSyntaxSourceCodeFor:.
	* RewriteTests.st: Add the TestRewrite class.
	* package.xml: Add the TestRewrite test to the testsuite.