stinst: Fix the sourcecode extraction in RBScanner>>#scanNumber
Fix two problems introduced by the RBNumberLiteralToken change.
The first one is to use >>#copyFrom:to: instead of re-positioning
the stream to fix an issue with the ConcatenatedStream and the
second is to pick the right stop for the token.

Code like '(Delay forSeconds: 3) wait' got re-formatted to
'(Delay forSeconds: 3)) wait'. The code that extracts the
sourcecode for the RBNumberLiteralToken should have operated on
the result of >>#previousStepPosition but instead the current
position of stream was used as an end.

Use self previousStepPosition to find the end of the number literal,
remove unused variables and remove the trimSeparators as we are now
using the right place to stop.

2013-02-17  Holger Hans Peter Freyther  <holger@freyther.de>

	* RBParser.st: Fix RBScanner>>#scanNumber.
	* RewriteTests.st: Add testcase for RBScanner>>#scanNumber.
	* package.xml: Add the new test to the testsuite.