stinst: Fix parsing of negated numbers with the new RBNumberLiteralToken
Negative numbers are not directly parsed in the RBScanner>>#scanNumber
but are made negative from within the RBParser>>#parseNegatedNumber. This
was done to help to differentiate a binary selector from the number. Add
a testcase for the formatting, make the number negative inside the number
literal token and prepend the $- to the source.

The start/stop of the token has been wrong before this commit and is still
wrong. This needs to be fixed in the future.