postgres: Work on formatting Smalltalk types for PostgreSQL
DBD-PostgreSQL had various issues with >>#executeWithAll: on
prepared statements with various datatypes. Create a testcase
and make it configurable like it is done for MySQL.

>>#executeWithAll: didn't allow to insert a Smalltalk nil
into the database. Update the code formatting the query to
have a NULL pointer instead of a string with the value "NULL".

The FieldConverter is used differently between MySQL and
SQLite/PostgreSQL. On MySQL it may be used to help to format
a query (but no string escaping is done so it remains a
dangerous function) and on the other implementations it is
used for formatting the parameters when executing a prepared
statement.

On PostgreSQL the extra quotes lead to issues importing data.
First of all the FieldConverter never called the existing write
functions for Booleans and Integers. This is corrected by adding
True/False to the generic lookup table and special casing
integers. Then >>#writeBoolean:on: and >>#writeInteger:on: of
PGFieldConverter do not write quotes on the stream.

The last fix is with printing DateTime. On PostgreSQL one can
specify if a DATETIME column includes a timezone or not. When
inserting a DateTime this information should not be included
as PostgreSQL will do the right thing. The comparison of the
offset should have been with Duration zero but we can simply
omit this code.

2013-05-20  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* configure.ac: Add --enable-postgres-tests.
	* tests/atlocal.in: Add variables for PostgreSQL tests.
	* tests/local.at: Ignore the output on stderr for AT_PACKAGE_TEST.
	* tests/testsuite.at: Add the DBD-PostgreSQL package.

2013-05-20  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* FieldConverter.st: Add conversion for Integer, fix
	conversion for Boolean, Integer and DateTime.
	* Statement.st: Modify >>#executeWithAll:
	* Tests.st: Add the PostgresTestCase class.
	* package.xml: Add Test.st and sunit section.