kernel: Fix references when copying methods and add new test package
When a compiled method is copied some literals (block and closures)
need to be updated as they have references to the old method. The
debug information also needs to be updated to point to the new
literals array. Implement this and add a new package that will use
SUnit test cases to test kernel classes.

2013-08-10  Gwenael Casaccio <gwenael.casaccio@gmail.com>

	* kernel/BlkClosure.st: Add >>#method: setter.
	* kernel/CompildMeth.st: Implement >>#copy and >>#deepCopy.
	* kernel/CompiledBlk.st: Add >>#method: setter.
	* kernel/MethodInfo.st: Add >>#debugInformation, >>#postCopy
	>>#fixDebugInformation:replaceWith: and >>#debugInfoReplace.
	* configure.ac: Introduce the Kernel-Tests package.
	* tests/testsuite.at: Add the Kernel-Tests package

2013-06-11  Gwenael Casaccio <gwenael.casaccio@gmail.com>

	 * kernel/CompiledMethodTests.st: Add tests for CompiledMethod