visualgst: Fix method selection in the hierarchy widget
Moving to "Hierarchy" and then moving to a parent- or subclass
one could see the available categories but it was not possible
to browse methods belonging to that class. This is because the
class selection was done using >>#selectedClass on the classWidget.

Change the code to use the classOrMeta of the GtkCategoryWidget
to determine the class to look at. The concept of >>#selectedClass
is still not consistent in this class. The class should be probably
split into a few smaller ones.

2014-02-01  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* StBrowser/GtkClassBrowserWidget.st: Use >>#classOrMeta
	of GtkCategoryWidget for the class.