xlat: Using BYTECODE_SIZE is not always the right thing
Use the difference between IP and IP0 to figure out how many bytes
were used for the bytecode of the send message. Add a small test
case and try to make it as stable as possible.

For message sends we remember the bytecode length as calculated
by the decoding of the bytecodes.

2014-02-10  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* xlat.c: Keep track of the bytecode length for sends.