libgst: Free memory to avoid memory leaks
sval was allocated with strdup but not freed when the obstack of
the compilation was released. Allocate the string inside the obstack
to avoid this issue. Remember if a fileName has been dynamically
allocated and then free it. Free a buffer even if it contains the
empty string, free the function name after the function has been
resolved.

2014-02-10  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* input.c: Free memory that has been dynamically allocated.
	* lex.c: Allocate string in the obstack.
	* prims.def: Free function name after usage.