netlink: Introduce a Linux/Netlink extension for GNU Smalltalk
The module only supports one query right now but it could and
should grow over time. The Transport of the OsmoSIP binds to the
unknown address but when addressing a UAS/Proxy it needs to fill
in a real IPv4 Address in the Contact header. This module can be
used for that.

  (Sockets.SocketAddress byName: 'remote') determineSourceAddress

2014-03-25  Holger Hans Peter Freyther  <holger@moiji-mobile.com>

	* Extensions.st: Add file.
	* Netlink.st: Add file.
	* netlink.c: Add file.
	* package.xml: Add file.