visualgst: Avoid live-lock on single-step debugging
Debugging "3 negated" and pressing "F7" could result in a
live-lock. While the dispatch routine of the mainloop was
ran the call-in process could wait on the "finalSemaphore"
and the "idle" process would be selected to execute as the
only process that remained runnable. This live-lock would
never be exited.

The underlying issue is that VisualGST is running in the
same image and can be impacted by the execution of the
debugged process. The only thing we can do right now is
to skip dangerous routines.

Currently the dangerous routines are related to the handling
of finalizers as the Glib/Gtk+ bindings create a lot of to
be finalized objects.

2014-05-05  Holger Hans Peter Freyther  <holger@freyther.de>

	* Debugger/GtkDebugger.st: Introduce >>#finishDangerousContexts
	and call it from >>#debugWith:.