gst-sunit: Spawn the event loop last to avoid crashes
It was possible that while we started to show the VisualGST
window GTK+ would call back into Smalltalk and we would run
the Glib dispatch while showing the window. This lead to a
timer expiring on a non-fully realized window and leading to
a crash inside GTK+/X11.

Run the event loop after we have intitialized all of the
VisualGST code.

2014-07-20  Holger Hans Peter Freyther  <holger@freyther.de>

	* scripts/Test.st: Spawn the event loop last.