libgst: Add built-in for Behavior>>#new and >>#new:
The addition of calling both basicNew/basicNew: followed by a
call to initialize introduced a noticable slowdown in an
allocation macro benchmark. By making the new >>new/>>new:
a primitive we reduce the cost.

The #initialize symbol is part of the builtin selectors so
we can not blindly allocate it during the construction of
the symbol table but will take the result of the built-in
selectors.

2014-08-02  Holger Hans Peter Freyther  <holger@freyther.de>

	* prims.def: Introduce VMpr_Behavior_newInitialize and
	VMpr_Behavior_newColonInitialize,
	* sym.h: Declare _gst_initialize_symbol.
	* sym.c: Define _gst_initialize_symbol during init and
	restore.

2014-08-02  Holger Hans Peter Freyther  <holger@freyther.de>

	* kernel/Builtins.st: Use the new built-in for >>#new
	and >>#new:.