libgst: Fix small amount of sleeps on OSX
create_timer is not available on OSX which means the
code will fall back to use setitimer. In case we want
to sleep a short amount of time we either cancel the
timeout or sleep a very long time.

* If nsTime < now we would sleep a very long time.
* If deltaMilli == 0 we could cancel the timer and
never wake up.

If the delta is smaller than what can be expressed in
milliseconds sleep the shortest time possible as it is
too late to cancel everything.

This could be easily reproduced using the following code
which would be unlikely to run to the end.

 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait
 (Delay forMilliseconds: 1) wait