libgst: Add alternative multiplication overflow check
Apple clang on OSX and the version on FreeBSD optimize the
multiplication check away. Clang introduced a family of
builtins to do the multiplication and check for the overflow
and GCC made the API usable. For clang we would need to know
if intptr_t is of type int, long int, long long int and
then use the smul, smull smulll.
Luckily clang is adopting the better interface and this is
what we are starting to use now. This means the new code
will be used on GCC5 (and later) and some future versions of
clang.

2015-11-07  Holger Hans Peter Freyther  <holger@freyther.de>

	* build-aux/overflow-builtins.m4: Add new macro.
	* configure.ac: Use GST_C_OVERFLOW_BUILTINS macro.

2015-11-07  Holger Hans Peter Freyther  <holger@freyther.de>

	* interp.inl: Add alternative mul_with_check implementation.