blox: Fix fall-out from calling >>#initialize as part of new
The code assumes that parentView is set before >>#initialize is
called and this assumption was broken by the change. In the long
run it might be best to call this routine >>#initializeView and
or integrate the setting of the parent view. The patch changes
the base class to restore the old behavior for creation.

Call >>#initialize on the Workspace so that the variableNames
tracker is properly initialized. Remove one unneeded call for
the STInST module.

2016-01-23  Holger Hans Peter Freyther  <holger@freyther.de>

	* BrowShell.st: Call >>#initialize.
	* DebugSupport.st: Do not call >>#initialize.
	* View.st: Use >>#basicNew to instantiate.

2016-01-23  Holger Hans Peter Freyther  <holger@freyther.de>

	* BloxBasic.st: Use >>#basicNew to create Gui items.