genbc: Deal with multiple definition errors
genbc is a special binary that is using two bison generated parsers and
a single lexer. The lexer and parser exchange data through the yylval
variable. Newer versions of the GNU linker detect the multiple
definitions.

Avoid this by mapping yychar and yynerrs to private state and yylval to
call a function to get the value.

Fixes:
/usr/bin/ld: genbc-impl.o:(.bss+0xc): multiple definition of `yychar'; genbc-decl.o:(.bss+0x4c): first defined here
/usr/bin/ld: genbc-impl.o:(.bss+0x10): multiple definition of `yylval'; genbc-decl.o:(.bss+0x50): first defined here
/usr/bin/ld: genbc-impl.o:(.bss+0x18): multiple definition of `yynerrs'; genbc-decl.o:(.bss+0x58): first defined here