mem: macos on arm64 does not allow write (and read) on executable pages
On macos with apple silicon mmap with PROT_EXEC|PROT_WRITE will fail.
This caused the heap code to fail. We do not expect to execute any of
these pages. So let's remove this bit.

The jit code will require some more changes to support RW^X on hardened
systems.

The change is enough to make basics of gst run on macos/arm64.